Prevent "field is immutable" error on Helm 3 upgrade
When attempting to upgrade openldap resources with Helm client v3,
OpenLDAP resources fail to upgrade after initial installation because
we DON'T supply a static cluster IP for these resources.

As described in the referenced discussion below, when a cluster ip
isn't supplied for this chart, the Helm 3 client sees the "live" state
of the resource as having the cluster assigned IP address, but our
requested state is to unset that IP address - IE the cluster IP value
is "".

This change aims to prevent this from happening by ommitting the
cluster ip value from the template altogether if no value is
supplied.

References
----------

- https://github.com/helm/helm/issues/6378#issuecomment-557746499

Signed-off-by: Aaron Kuehler <aaron.kuehler@gmail.com>